+++ src/commands/force/lightning/app/create.ts
@@ -4,17 +4,17 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { flags, SfdxCommand } from '@salesforce/command';
+import { flags } from '@salesforce/command';
 import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import * as path from 'path';
 import LightningAppGenerator from '../../../../generators/lightningAppGenerator';
-import { CreateUtil, SfdxCommandBase } from '../../../../utils';
+import { CreateUtil, TemplateCommand } from '../../../../utils';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('salesforcedx-templates', 'messages');
 const lightningAppFileSuffix = /.app$/;
-export default class LightningApp extends SfdxCommandBase {
+export default class LightningApp extends TemplateCommand {
   public static examples = [
     '$ sfdx force:lightning:app:create -n myapp',
     '$ sfdx force:lightning:app:create -n myapp -d aura'
