+++ src/utils/index.ts
@@ -2,4 +2,4 @@ export { CreateUtil } from './createUtil';
 export { Log } from './logger';
 export { ForceGeneratorAdapter } from './adapter';
 export { Answers } from './types';
-export { SfdxCommandBase } from './SfdxCommandBase';
+export { TemplateCommand } from './templateCommand';
