+++ src/utils/templateCommand.ts
@@ -7,42 +7,59 @@
 
 import { SfdxCommand } from '@salesforce/command';
 import { Messages } from '@salesforce/core';
-
+import { AnyJson } from '@salesforce/ts-types';
 // tslint:disable-next-line:no-var-requires
-const generator = require('yeoman-generator');
+const yeomanGenerator = require('yeoman-generator');
 import { ForceGeneratorAdapter } from './adapter';
 import * as path from 'path';
-import { CreateUtil } from './createUtil';
+import { CreateOutput } from './types';
 // tslint:disable-next-line:no-var-requires
 const yeoman = require('yeoman-environment');
 const messages = Messages.loadMessages('salesforcedx-templates', 'messages');
 
-export class SfdxCommandBase extends SfdxCommand {
+export abstract class TemplateCommand extends SfdxCommand {
   // tslint:disable-next-line:no-any
-  run(): Promise<any> {
-    throw new Error('Method not implemented.');
-  }
+  abstract run(): Promise<AnyJson>;
 
-  public async runGenerator(generatorname: typeof generator) {
+  public async runGenerator(generator: typeof yeomanGenerator) {
     // tslint:disable-next-line:no-unused-expression
     if (!this.flags.apiversion) {
-      this.flags.apiversion = CreateUtil.getDefaultApiVersion();
+      this.flags.apiversion = this.getDefaultApiVersion();
     }
 
     const adapter = new ForceGeneratorAdapter();
     const env = yeoman.createEnv(undefined, undefined, adapter);
-    env.registerStub(generatorname, 'generator');
+    env.registerStub(generator, 'generator');
 
     const result = await env.run('generator', this.flags);
     const targetDir = path.resolve(this.flags.outputdir);
 
     // tslint:disable-next-line:no-unused-expression
     if (this.flags.json) {
-      return CreateUtil.buildJson(adapter, targetDir);
+      return this.buildJson(adapter, targetDir);
     } else {
       this.log(messages.getMessage('targetDirOutput', [targetDir]));
       this.log(adapter.log.getOutput());
       return result;
     }
+  }
+
+  public getDefaultApiVersion(): string {
+    const versionTrimmed = require('../../package.json').version.trim();
+    return `${versionTrimmed.split('.')[0]}.0`;
+  }
+
+  public buildJson(
+    adapter: ForceGeneratorAdapter,
+    targetDir: string
+  ): CreateOutput {
+    const cleanOutput = adapter.log.getCleanOutput();
+    const rawOutput = `target dir = ${targetDir}\n${adapter.log.getOutput()}`;
+    const output = {
+      outputDir: targetDir,
+      created: cleanOutput,
+      rawOutput
+    };
+    return output;
   }
 }
