+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "./node_modules/@salesforce/dev-config/tsconfig",
   "compilerOptions": {
     "module": "commonjs",
     "target": "es6",
